Get legacy data version 1

Get legacy data version 1

Usage

get_data_v1(
  type = c("complete", "timeseries"),
  pcode = c("s1", "s2", "s3", "s4", "s5", "s6", "s7", "s11", "s12", "s13"),
  notes = TRUE,
  directory = NULL,
  tidy = FALSE
)

get_data_v1_complete(notes = TRUE, directory = NULL)

get_data_v1_timeseries(
  pcode = c("s1", "s2", "s3", "s4", "s5", "s6", "s7", "s11", "s12", "s13"),
  directory = NULL,
  tidy = FALSE
)

Arguments

type

A character value for type of dataset to retrieve. Either "complete" for the full dataset or "timeseries" for the disaggregated policy indicators datasets. Default is "complete".

pcode

A character value for policy indicator code to retrieve if type is "timeseries". Ignored if type is "complete". Can be one of "s1", "s2", "s3", "s4", "s5", "s6", "s7", "s8", "s9", "s10", "s11", "s12", or "s13".

notes

Logical. Should "complete" dataset be retrieved with notes or not. Default is TRUE. Ignored if type is "timeseries".

directory

If NULL (default), required dataset is read into current environment. If specified, required dataset is downloaded into this directory with the filename given by the OxCGRT.

tidy

Logical. Should "timeseries" dataset be converted to tidy format? Default is FALSE.

Value

If directory is NULL, a tibble (or a list of two tibbles) of the specified dataset. Otherwise, a file path/s to the downloaded file/s.

Examples

get_data_v1_complete()
#> # A tibble: 17,282 × 39
#>    CountryName CountryCode Date       `S1_School closing` S1_IsGeneral S1_Notes
#>    <chr>       <chr>       <date>                   <dbl>        <dbl> <chr>   
#>  1 Aruba       ABW         2020-01-01                   0           NA NA      
#>  2 Aruba       ABW         2020-01-02                   0           NA NA      
#>  3 Aruba       ABW         2020-01-03                   0           NA NA      
#>  4 Aruba       ABW         2020-01-04                   0           NA NA      
#>  5 Aruba       ABW         2020-01-05                   0           NA NA      
#>  6 Aruba       ABW         2020-01-06                   0           NA NA      
#>  7 Aruba       ABW         2020-01-07                   0           NA NA      
#>  8 Aruba       ABW         2020-01-08                   0           NA NA      
#>  9 Aruba       ABW         2020-01-09                   0           NA NA      
#> 10 Aruba       ABW         2020-01-10                   0           NA NA      
#> # ℹ 17,272 more rows
#> # ℹ 33 more variables: `S2_Workplace closing` <dbl>, S2_IsGeneral <dbl>,
#> #   S2_Notes <chr>, `S3_Cancel public events` <dbl>, S3_IsGeneral <dbl>,
#> #   S3_Notes <chr>, `S4_Close public transport` <dbl>, S4_IsGeneral <dbl>,
#> #   S4_Notes <chr>, `S5_Public information campaigns` <dbl>,
#> #   S5_IsGeneral <dbl>, S5_Notes <chr>,
#> #   `S6_Restrictions on internal movement` <dbl>, S6_IsGeneral <dbl>, …
get_data_v1_timeseries()
#> $s1_isgeneral
#> # A tibble: 153 × 116
#>    CountryName       CountryCode `01jan2020` `02jan2020` `03jan2020` `04jan2020`
#>    <chr>             <chr>       <lgl>       <lgl>       <lgl>       <lgl>      
#>  1 Aruba             ABW         NA          NA          NA          NA         
#>  2 Afghanistan       AFG         NA          NA          NA          NA         
#>  3 Angola            AGO         NA          NA          NA          NA         
#>  4 Albania           ALB         NA          NA          NA          NA         
#>  5 Andorra           AND         NA          NA          NA          NA         
#>  6 United Arab Emir… ARE         NA          NA          NA          NA         
#>  7 Argentina         ARG         NA          NA          NA          NA         
#>  8 Australia         AUS         NA          NA          NA          NA         
#>  9 Austria           AUT         NA          NA          NA          NA         
#> 10 Azerbaijan        AZE         NA          NA          NA          NA         
#> # ℹ 143 more rows
#> # ℹ 110 more variables: `05jan2020` <lgl>, `06jan2020` <lgl>,
#> #   `07jan2020` <lgl>, `08jan2020` <lgl>, `09jan2020` <lgl>, `10jan2020` <lgl>,
#> #   `11jan2020` <lgl>, `12jan2020` <lgl>, `13jan2020` <lgl>, `14jan2020` <lgl>,
#> #   `15jan2020` <lgl>, `16jan2020` <lgl>, `17jan2020` <lgl>, `18jan2020` <lgl>,
#> #   `19jan2020` <lgl>, `20jan2020` <lgl>, `21jan2020` <lgl>, `22jan2020` <lgl>,
#> #   `23jan2020` <lgl>, `24jan2020` <dbl>, `25jan2020` <dbl>, …
#> 
#> $s1_schoolclosing
#> # A tibble: 153 × 116
#>    CountryName       CountryCode `01jan2020` `02jan2020` `03jan2020` `04jan2020`
#>    <chr>             <chr>             <dbl>       <dbl>       <dbl>       <dbl>
#>  1 Aruba             ABW                   0           0           0           0
#>  2 Afghanistan       AFG                   0           0           0           0
#>  3 Angola            AGO                   0           0           0           0
#>  4 Albania           ALB                   0           0           0           0
#>  5 Andorra           AND                   0           0           0           0
#>  6 United Arab Emir… ARE                   0           0           0           0
#>  7 Argentina         ARG                   0           0           0           0
#>  8 Australia         AUS                   0           0           0           0
#>  9 Austria           AUT                   0           0           0           0
#> 10 Azerbaijan        AZE                   0           0           0           0
#> # ℹ 143 more rows
#> # ℹ 110 more variables: `05jan2020` <dbl>, `06jan2020` <dbl>,
#> #   `07jan2020` <dbl>, `08jan2020` <dbl>, `09jan2020` <dbl>, `10jan2020` <dbl>,
#> #   `11jan2020` <dbl>, `12jan2020` <dbl>, `13jan2020` <dbl>, `14jan2020` <dbl>,
#> #   `15jan2020` <dbl>, `16jan2020` <dbl>, `17jan2020` <dbl>, `18jan2020` <dbl>,
#> #   `19jan2020` <dbl>, `20jan2020` <dbl>, `21jan2020` <dbl>, `22jan2020` <dbl>,
#> #   `23jan2020` <dbl>, `24jan2020` <dbl>, `25jan2020` <dbl>, …
#> 
get_data_v1()
#> # A tibble: 17,282 × 39
#>    CountryName CountryCode Date       `S1_School closing` S1_IsGeneral S1_Notes
#>    <chr>       <chr>       <date>                   <dbl>        <dbl> <chr>   
#>  1 Aruba       ABW         2020-01-01                   0           NA NA      
#>  2 Aruba       ABW         2020-01-02                   0           NA NA      
#>  3 Aruba       ABW         2020-01-03                   0           NA NA      
#>  4 Aruba       ABW         2020-01-04                   0           NA NA      
#>  5 Aruba       ABW         2020-01-05                   0           NA NA      
#>  6 Aruba       ABW         2020-01-06                   0           NA NA      
#>  7 Aruba       ABW         2020-01-07                   0           NA NA      
#>  8 Aruba       ABW         2020-01-08                   0           NA NA      
#>  9 Aruba       ABW         2020-01-09                   0           NA NA      
#> 10 Aruba       ABW         2020-01-10                   0           NA NA      
#> # ℹ 17,272 more rows
#> # ℹ 33 more variables: `S2_Workplace closing` <dbl>, S2_IsGeneral <dbl>,
#> #   S2_Notes <chr>, `S3_Cancel public events` <dbl>, S3_IsGeneral <dbl>,
#> #   S3_Notes <chr>, `S4_Close public transport` <dbl>, S4_IsGeneral <dbl>,
#> #   S4_Notes <chr>, `S5_Public information campaigns` <dbl>,
#> #   S5_IsGeneral <dbl>, S5_Notes <chr>,
#> #   `S6_Restrictions on internal movement` <dbl>, S6_IsGeneral <dbl>, …