Calculate an OxCGRT index or indices

Calculate an OxCGRT index or indices

calculate_index(df, codes, tolerance)

calculate_gov_response(df)

calculate_containment_health(df)

calculate_stringency(df)

calculate_economic_support(df)

calculate_indices(df)

Arguments

df
A data.frame produced by a call to calculate_subindices().

codes
A vector of policy type codes to use for the index calculation.

tolerance
An integer specifying the number of missing values above which index will not be calculated and reported.

Value

A numeric value for mean subindex scores of specified policy types. For calculate_indices(), a tibble calculated OxCGRT indices

Author

Ernest Guevarra

Examples

## Get policy actions data for Afghanistan on 1 September 2020
x <- get_data(json = get_json_actions(ccode = "AFG",
                                      from = NULL,
                                      to = "2020-09-01"))

## Calculate OxCGRT subindices
y <- calculate_subindices(df = x$policyActions)

## Calculate OxCGRT index
calculate_index(df = y,
                codes = c(paste("C", 1:8, sep = ""),
                          paste("E", 1:2, sep = ""),
                          paste("H", 1:3, sep = ""), "H6"),
                tolerance = 1)

#> [1] 33.33333

## Calculate OxCGRT government response index
calculate_gov_response(df = y)

#> [1] 33.33333

## Calculate OxCGRT containment and health index
calculate_containment_health(df = y)

#> [1] 38.88889

## Calculate OxCGRT stringency index
calculate_stringency(df = y)

#> [1] 21.2963

## Calculate OxCGRT economic support index
calculate_economic_support(df = y)

#> [1] 0

## Calculate all OxCGRT indices
calculate_indices(df = y)

#> # A tibble: 4 x 2
#>   index                        values
#>   <chr>                         <dbl>
#> 1 Government Response Index      33.3
#> 2 Containment and Health Index   38.9
#> 3 Stringency Index               21.3
#> 4 Economic Support Index          0